Optimize ChatConsole relayout calls
Don't relayout once per message, just once.

It was taking hell of a time for a long list of messages
(and it's still taking too much time).

I ran into an issue because some messages are duplicated in the
_chatMessages array. The fix in ChatMessage.as:287 prevents more
issues but we should probably figure out the source of it.